% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embellish_cox.R
\name{embellish_cox}
\alias{embellish_cox}
\title{Modify the results of COX analysis}
\usage{
embellish_cox(data = NULL, Characteristics = NULL)
}
\arguments{
\item{data}{Mul_cox results}

\item{Characteristics}{Characteristics}
}
\value{
Cox results
}
\description{
Modify the results of COX analysis
}
\examples{
Mul_res = embellish_cox(data = Mul_res, Characteristics = Characteristics)

}
