% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format_pvalue.R
\name{format_pvalue}
\alias{format_pvalue}
\title{Format pvalue}
\usage{
format_pvalue(p = NULL, accuracy = 0.01, flag = "#")
}
\arguments{
\item{p}{p value}

\item{accuracy}{cutoff for Scientific notation}

\item{flag}{Whether to keep the trailing 0}
}
\value{
Formatted p value
}
\description{
Format pvalue
}
\examples{
format_pvalue(p = 0.05, accuracy = 0.01)
format_pvalue(p = 0.005, accuracy = 0.01)
}
