% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_attr.R
\name{Get_attr}
\alias{Get_attr}
\title{Get attribution}
\usage{
Get_attr(col = NULL, print = T)
}
\arguments{
\item{col}{column of a dataframe}

\item{print}{Whether or not to print}
}
\value{
Arranged statistical results.
}
\description{
Get attribution
}
\examples{
Get_attr(iris$Species)
}
