% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cox_analysis.R
\name{Mul_cox}
\alias{Mul_cox}
\title{Characteristics' = x,
CI5' = CI5,
CI95' = CI95,}
\usage{
Mul_cox(
  variables = NULL,
  data = NULL,
  p_cut = 0.25,
  censored = NULL,
  OS = "OS",
  OS_time = "OS_time",
  included_uni = T
)
}
\description{
Characteristics' = x,
CI5' = CI5,
CI95' = CI95,
}
