% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stat_col.R
\name{Stat_col}
\alias{Stat_col}
\title{Get statistical results of column}
\usage{
Stat_col(col = NULL)
}
\arguments{
\item{col}{column of a dataframe}
}
\value{
Arranged statistical results, frequency, percent.
}
\description{
Get statistical results of column
}
\examples{
Stat_col(iris$Species)
}
