% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cox_analysis.R
\name{Uni_cox}
\alias{Uni_cox}
\title{Cox Analysis}
\usage{
Uni_cox(
  para = NULL,
  data = NULL,
  OS = "OS",
  OS_time = "OS_time",
  censored = 60,
  return_raw = F
)
}
\arguments{
\item{para}{Variable for analysis}

\item{data}{Clinial Matrix}

\item{OS}{Column names representing survival status}

\item{OS_time}{Column names representing survival time}

\item{censored}{OS time cutoff; 60 is recommended}
}
\value{
Cox univariate and multivariate results
}
\description{
Cox Analysis
}
